les controllers

+++ src/controllers/ChambreController.js
@@ -1,9 +1,40 @@
-import prisma from "../config/prisma.js";
+import { PrismaClient } from "@prisma/client";
+
+const prisma = new PrismaClient();
 
 class ChambreController {
   static async creerChambre(req, res) {
-    const { numeroChambre, prix, typeId, utilisateurId } = req.body;
     try {
+      const { numeroChambre, prix, typeId, utilisateurId } = req.body;
+
+      const chambreExistant = await prisma.chambre.findUnique({
+        where: { numeroChambre }
+      });
+
+      if (chambreExistant) {
+        return res
+          .status(400)
+          .json({ message: "Le  numéro du chambre est déjà utilisé" });
+      }
+
+      const typeExiste = await prisma.typeChambre.findUnique({
+        where: { id: typeId }
+      });
+
+      if (!typeExiste) {
+        return res
+          .status(404)
+          .json({ message: "Le type de chambre n'existe pas" });
+      }
+
+      const utilisateurExiste = await prisma.utilisateur.findUnique({
+        where: { id: utilisateurId }
+      });
+
+      if (!utilisateurExiste) {
+        return res.status(404).json({ message: "L'utilisateur n'existe pas" });
+      }
+
       const nouvelleChambre = await prisma.chambre.create({
         data: {
           numeroChambre,
@@ -60,6 +91,39 @@ class ChambreController {
     const { id } = req.params;
     const { numeroChambre, prix, typeId, utilisateurId } = req.body;
     try {
+      const chambre = await prisma.chambre.findUnique({
+        where: { id: parseInt(id) }
+      });
+
+      if (!chambre) {
+        return res.status(404).json({ message: "Chambre non trouvé" });
+      }
+      const chambreExistant = await prisma.chambre.findUnique({
+        where: { numeroChambre }
+      });
+
+      if (chambreExistant) {
+        return res.status(400).json({ message: "Cet numéro est déjà utilisé" });
+      }
+
+      const typeExiste = await prisma.typeChambre.findUnique({
+        where: { id: typeId }
+      });
+
+      if (!typeExiste) {
+        return res
+          .status(404)
+          .json({ message: "Le type de chambre n'existe pas" });
+      }
+
+      const utilisateurExiste = await prisma.utilisateur.findUnique({
+        where: { id: utilisateurId }
+      });
+
+      if (!utilisateurExiste) {
+        return res.status(404).json({ message: "L'utilisateur n'existe pas" });
+      }
+
       const chambreMisAJour = await prisma.chambre.update({
         where: { id: parseInt(id) },
         data: {
@@ -82,10 +146,17 @@ class ChambreController {
   static async supprimerChambre(req, res) {
     const { id } = req.params;
     try {
+      const chambre = await prisma.chambre.findUnique({
+        where: { id: parseInt(id) }
+      });
+
+      if (!chambre) {
+        return res.status(404).json({ message: "Chambre non trouvé" });
+      }
       await prisma.chambre.delete({
         where: { id: parseInt(id) }
       });
-      return res.json({ message: "Utilisateur supprimer avec sucess" });
+      return res.json({ message: "Chambre supprimer avec sucess" });
     } catch (error) {
       console.error(error);
       return res.status(500).json({
